Update docker container with smaller initial image
