+++ pyproject.toml
@@ -9,7 +9,7 @@ readme = "README.md"
 python = "^3.11"
 flask = "^2.2.2"
 pytest = "^7.2.0"
-locust = "^2.13.0"
+psutil = "^5.9.4"
 coverage = "^7.0.3"
 marshmallow = "^3.19.0"
 inflection = "^0.5.1"
