+++ tests/configs/locust.conf
@@ -5,4 +5,4 @@ spawn-rate = 50
 run-time = 60s
 expect-workers = 1
 only-summary = true
-locustfile = /mnt/locust/seabattle_test.py
+locustfile = /mnt/locust/seabattle.py
